module5-task1 - Модуляция
